src/dashboard-app/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.3
plotly==5.18.0
google-cloud-bigquery==3.14.0
db-dtypes==1.2.0
